display content depending on auth state

+++ app.js
@@ -8,43 +8,49 @@ document.addEventListener('DOMContentLoaded', ()=>{
 })
 
 // create elements and render to DOM
-function renderCafe(doc) {
-    let li = document.createElement('li')
-    let name = document.createElement('span')
-    let city = document.createElement('span')
-    let cross = document.createElement('div')
+function renderCafe(data) {
+    if(data.length > 0) {
+        data.forEach(doc =>{
+            let li = document.createElement('li')
+            let name = document.createElement('span')
+            let city = document.createElement('span')
+            let cross = document.createElement('div')
+        
+            li.setAttribute('data-id', doc.id)
+            name.textContent = doc.data().name
+            city.textContent = doc.data().city
+            cross.textContent = 'x'
+        
+            li.appendChild(name)
+            li.appendChild(city)
+            li.appendChild(cross)
+        
+            cafeList.appendChild(li)
+        
+            cross.addEventListener('click', (e)=>{
+                e.stopPropagation()
+                let id = e.target.parentElement.getAttribute('data-id')
+                db.collection('cafe').doc(id).delete()
+                cafeList.innerHTML = ''
+                init()
+            })
+        
+            name.addEventListener('click', (e)=>{
+                const val = e.target.innerText
+                e.target.innerHTML = `<form onsubmit="editData(${e.target.parentElement.getAttribute('data-id')})"><input type="text" value='${val}' /></form>`
+            })
+        })
+    }
+    else {
+        cafeList.innerHTML = '<h5>you cant see this :)</h5>'
+    }
 
-    li.setAttribute('data-id', doc.id)
-    name.textContent = doc.data().name
-    city.textContent = doc.data().city
-    cross.textContent = 'x'
-
-    li.appendChild(name)
-    li.appendChild(city)
-    li.appendChild(cross)
-
-    cafeList.appendChild(li)
-
-    cross.addEventListener('click', (e)=>{
-        e.stopPropagation()
-        let id = e.target.parentElement.getAttribute('data-id')
-        db.collection('cafe').doc(id).delete()
-        cafeList.innerHTML = ''
-        init()
-    })
-
-    name.addEventListener('click', (e)=>{
-        const val = e.target.innerText
-        e.target.innerHTML = `<form onsubmit="editData(${e.target.parentElement.getAttribute('data-id')})"><input type="text" value='${val}' /></form>`
-    })
 }
 
 function init() {
     // get data using snapshot
     db.collection('cafe').get().then((snapshot)=>{
-        snapshot.docs.forEach(doc => {
-            renderCafe(doc);
-        });
+            renderCafe(snapshot.docs);
     })
 }
 
@@ -90,4 +96,4 @@ function editData(id, n) {
     })
 }
 
-window.addEventListener('load', ()=> init())
+// window.addEventListener('load', ()=> init())
